+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # ![Logo](https://github.com/WebTools-NG/WebTools-NG/blob/master/src/assets/WebTools-48x48.png) WebTools-ng Change log
 
-## V1.1.2 ( Not released yet )
+## V1.1.2 ( 20220925 )
 
 **Note**: In this version, the following is disabled: